+++ scope.go
@@ -2,6 +2,7 @@ package sentry
 
 import (
 	"reflect"
+	"sync"
 	"time"
 )
 
@@ -21,6 +22,7 @@ import (
 // Note that the scope can only be modified but not inspected.
 // Only the client can use the scope to extract information currently.
 type Scope struct {
+	sync.RWMutex
 	breadcrumbs     []*Breadcrumb
 	user            User
 	tags            map[string]string
@@ -52,6 +54,9 @@ func (scope *Scope) AddBreadcrumb(breadcrumb *Breadcrumb, limit int) {
 		breadcrumb.Timestamp = time.Now().Unix()
 	}
 
+	scope.Lock()
+	defer scope.Unlock()
+
 	breadcrumbs := append(scope.breadcrumbs, breadcrumb)
 	if len(breadcrumbs) > limit {
 		scope.breadcrumbs = breadcrumbs[1 : limit+1]
@@ -77,11 +82,17 @@ func (scope *Scope) SetRequest(request Request) {
 
 // SetTag adds a tag to the current scope.
 func (scope *Scope) SetTag(key, value string) {
+	scope.Lock()
+	defer scope.Unlock()
+
 	scope.tags[key] = value
 }
 
 // SetTags assigns multiple tags to the current scope.
 func (scope *Scope) SetTags(tags map[string]string) {
+	scope.Lock()
+	defer scope.Unlock()
+
 	for k, v := range tags {
 		scope.tags[k] = v
 	}
@@ -89,16 +100,25 @@ func (scope *Scope) SetTags(tags map[string]string) {
 
 // RemoveTag removes a tag from the current scope.
 func (scope *Scope) RemoveTag(key string) {
+	scope.Lock()
+	defer scope.Unlock()
+
 	delete(scope.tags, key)
 }
 
 // SetContext adds a context to the current scope.
 func (scope *Scope) SetContext(key string, value interface{}) {
+	scope.Lock()
+	defer scope.Unlock()
+
 	scope.contexts[key] = value
 }
 
 // SetContexts assigns multiple contexts to the current scope.
 func (scope *Scope) SetContexts(contexts map[string]interface{}) {
+	scope.Lock()
+	defer scope.Unlock()
+
 	for k, v := range contexts {
 		scope.contexts[k] = v
 	}
@@ -106,16 +126,25 @@ func (scope *Scope) SetContexts(contexts map[string]interface{}) {
 
 // RemoveContext removes a context from the current scope.
 func (scope *Scope) RemoveContext(key string) {
+	scope.Lock()
+	defer scope.Unlock()
+
 	delete(scope.contexts, key)
 }
 
 // SetExtra adds an extra to the current scope.
 func (scope *Scope) SetExtra(key string, value interface{}) {
+	scope.Lock()
+	defer scope.Unlock()
+
 	scope.extra[key] = value
 }
 
 // SetExtras assigns multiple extras to the current scope.
 func (scope *Scope) SetExtras(extra map[string]interface{}) {
+	scope.Lock()
+	defer scope.Unlock()
+
 	for k, v := range extra {
 		scope.extra[k] = v
 	}
@@ -123,6 +152,9 @@ func (scope *Scope) SetExtras(extra map[string]interface{}) {
 
 // RemoveExtra removes a extra from the current scope.
 func (scope *Scope) RemoveExtra(key string) {
+	scope.Lock()
+	defer scope.Unlock()
+
 	delete(scope.extra, key)
 }
 
@@ -143,6 +175,9 @@ func (scope *Scope) SetTransaction(transactionName string) {
 
 // Clone returns a copy of the current scope with all data copied over.
 func (scope *Scope) Clone() *Scope {
+	scope.RLock()
+	defer scope.RUnlock()
+
 	clone := NewScope()
 	clone.user = scope.user
 	clone.breadcrumbs = make([]*Breadcrumb, len(scope.breadcrumbs))
@@ -161,6 +196,7 @@ func (scope *Scope) Clone() *Scope {
 	clone.level = scope.level
 	clone.transaction = scope.transaction
 	clone.request = scope.request
+
 	return clone
 }
 
@@ -171,11 +207,17 @@ func (scope *Scope) Clear() {
 
 // AddEventProcessor adds an event processor to the current scope.
 func (scope *Scope) AddEventProcessor(processor EventProcessor) {
+	scope.Lock()
+	defer scope.Unlock()
+
 	scope.eventProcessors = append(scope.eventProcessors, processor)
 }
 
 // ApplyToEvent takes the data from the current scope and attaches it to the event.
 func (scope *Scope) ApplyToEvent(event *Event, hint *EventHint) *Event {
+	scope.RLock()
+	defer scope.RUnlock()
+
 	if len(scope.breadcrumbs) > 0 {
 		if event.Breadcrumbs == nil {
 			event.Breadcrumbs = []*Breadcrumb{}